Refactor for clarity
Push all the oauth2 client stuff into a seperate file
Main method thus becomes a lot more streamlined
